content-moderation/moderator-model/requirements.txt
torch
numpy
pydantic
daily-python
transformers
pillow